web/requirements.txt
Django==1.8.8
gunicorn==19.3.0
psycopg2==2.6
django-redis-cache==1.6.5
redis==2.10.3
django-bower==5.0.4
django-pipeline==1.5.1
django-rest-auth==0.6.0
djangorestframework==3.3.2
futures==3.0.3
six==1.10.0
wheel==0.24.0
pillow==3.1.0
tinytag
git+https://github.com/pando85/django-sendfile
git+https://github.com/pando85/audiotranscode@0.3
